Added a socket.on to recieve data from leaderboard. Implemented to work with existing table generation.
