--- sampleleaderboarddata.js
@@ -1,42 +1,0 @@
-export const sampleleaderboarddata = [
-    {
-        name: "Sampleuser1",
-        points: 100000
-    },
-    {
-        name: "Sampleuser2",
-        points: 80000
-    },
-    {
-        name: "Sampleuser3",
-        points: 70000
-    },
-    {
-        name: "Sampleuser4",
-        points: 65000
-    },
-    {
-        name: "Sampleuser5",
-        points: 52000
-    },
-    {
-        name: "Sampleuser6",
-        points: 46000
-    },
-    {
-        name: "Sampleuser7",
-        points: 40000
-    },
-    {
-        name: "Sampleuser8",
-        points: 33000
-    },
-    {
-        name: "Sampleuser9",
-        points: 21000
-    },
-    {
-        name: "Sampleuser10",
-        points: 14000
-    },
-]
